+++ src/client/screens/OrganizationScreen.tsx
@@ -5,17 +5,22 @@ import { useParams } from "react-router-dom";
 import { Box, Button, Flex, Heading, Image, Link, jsx } from "theme-ui";
 
 import { organizationFetch } from "../actions/organization";
+import { joinOrganization, leaveOrganization } from "../actions/organizationJoin";
 import { userFetch } from "../actions/user";
 import { getJWT } from "../jwt";
 import { State } from "../reducers";
 import { OrganizationState } from "../reducers/organization";
+import { ProjectState } from "../reducers/project";
 import { UserState } from "../reducers/user";
 import store from "../store";
 import SiteHeader from "../components/SiteHeader";
 import Icon from "../components/Icon";
-
+import { IOrganization, IUser } from "../../shared/entities";
+import { showCopyMapModal } from "../actions/districtDrawing";
+import JoinOrganizationModal from "../components/JoinOrganizationModal";
 interface StateProps {
   readonly organization: OrganizationState;
+  readonly project: ProjectState;
   readonly user: UserState;
 }
 
@@ -50,9 +55,22 @@ const style = {
   }
 } as const;
 
-const OrganizationScreen = ({ organization, user }: StateProps) => {
+const OrganizationScreen = ({ organization, project, user }: StateProps) => {
   const { organizationSlug } = useParams();
   const isLoggedIn = getJWT() !== null;
+  const userInOrg =
+    "resource" in user &&
+    user.resource &&
+    "resource" in organization &&
+    organization.resource &&
+    checkIfUserInOrg(organization.resource, user.resource);
+
+  useEffect(() => {
+    "resource" in user &&
+      user.resource &&
+      project.showCopyMapModal &&
+      store.dispatch(showCopyMapModal(false));
+  }, [user, project.showCopyMapModal]);
 
   useEffect(() => {
     isLoggedIn && store.dispatch(userFetch());
@@ -61,6 +79,29 @@ const OrganizationScreen = ({ organization, user }: StateProps) => {
   useEffect(() => {
     store.dispatch(organizationFetch(organizationSlug));
   }, [organizationSlug]);
+
+  function joinOrg() {
+    "resource" in user &&
+      user.resource &&
+      store.dispatch(joinOrganization({ organization: organizationSlug, user: user.resource.id }));
+  }
+
+  function signupAndJoinOrg() {
+    store.dispatch(showCopyMapModal(true));
+  }
+
+  function leaveOrg() {
+    "resource" in user &&
+      user.resource &&
+      store.dispatch(leaveOrganization({ organization: organizationSlug, user: user.resource.id }));
+  }
+
+  function checkIfUserInOrg(org: IOrganization, user: IUser) {
+    const userExists = org.users.filter(u => {
+      return u.id === user.id;
+    });
+    return userExists.length > 0;
+  }
 
   return (
     <Flex sx={{ flexDirection: "column" }}>
@@ -89,26 +130,39 @@ const OrganizationScreen = ({ organization, user }: StateProps) => {
                     </Link>
                   </Box>
                 )}
-                {/* TODO #226: Show actual number of followers/members */}
                 <Box as="span" sx={style.item}>
-                  <Icon name="tools" /> 0 builders
+                  <Icon name="tools" /> {organization.resource.users.length} builders
                 </Box>
               </Box>
               {organization.resource.description && <Box>{organization.resource.description}</Box>}
             </Flex>
-            <Flex sx={{ flexDirection: "column", flex: "none" }}>
-              <Button disabled={true} sx={style.join}>
-                Join organization
-              </Button>
-              <Box sx={style.joinText}>
-                Join to start making district maps with this organization
-              </Box>
-            </Flex>
+            {userInOrg ? (
+              <Flex sx={{ flexDirection: "column", flex: "none" }} onClick={leaveOrg}>
+                <Button sx={style.join}>Leave organization</Button>
+              </Flex>
+            ) : "resource" in user && user.resource ? (
+              <Flex sx={{ flexDirection: "column", flex: "none" }} onClick={joinOrg}>
+                <Button sx={style.join}>Join organization</Button>
+                <Box sx={style.joinText}>
+                  Join to start making district maps with this organization
+                </Box>
+              </Flex>
+            ) : (
+              <Flex sx={{ flexDirection: "column", flex: "none" }} onClick={signupAndJoinOrg}>
+                <Button sx={style.join}>Join organization</Button>
+                <Box sx={style.joinText}>
+                  Register for an account to start making district maps with this organization
+                </Box>
+              </Flex>
+            )}
           </Flex>
         ) : (
           <Box>Loading...</Box>
         )}
       </Flex>
+      {"resource" in organization && organization.resource && (
+        <JoinOrganizationModal organization={organization.resource} />
+      )}
     </Flex>
   );
 };
@@ -116,6 +170,7 @@ const OrganizationScreen = ({ organization, user }: StateProps) => {
 function mapStateToProps(state: State): StateProps {
   return {
     organization: state.organization,
+    project: state.project,
     user: state.user
   };
 }