+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Duplicate a map from home screen [#572](https://github.com/PublicMapping/districtbuilder/pull/572)
 - Add script to load region configs [#575](https://github.com/PublicMapping/districtbuilder/pull/575)
 - Store chamber reference on project entity [#576](https://github.com/PublicMapping/districtbuilder/pull/576)
+- Allow users to join and leave an organization [#226](https://github.com/PublicMapping/districtbuilder/pull/578)
 
 ### Changed
 - Geounit label made more generic to support Dane County wards [#573](https://github.com/PublicMapping/districtbuilder/pull/573)